+++ docs/pages/components/datepicker/Datepicker.vue
@@ -24,6 +24,13 @@
             <p>You can add your custom header to the datepicker.</p>
         </Example>
 
+        <Example :component="ExMonth" :code="ExMonthCode" title="Month picker" vertical>
+            <div class="tags has-addons">
+                <span class="tag is-success">New!</span>
+                <span class="tag is-info">0.7.7</span>
+            </div>
+        </Example>
+
         <Example :component="ExProgrammatically" :code="ExProgrammaticallyCode" title="Programmatically opening" vertical>
             <div class="tags has-addons">
                 <span class="tag is-success">New!</span>
@@ -64,6 +71,9 @@
     import ExProgrammatically from './examples/ExProgrammatically'
     import ExProgrammaticallyCode from '!!raw-loader!./examples/ExProgrammatically'
 
+    import ExMonth from './examples/ExMonth'
+    import ExMonthCode from '!!raw-loader!./examples/ExMonth'
+
     import ExInline from './examples/ExInline'
     import ExInlineCode from '!!raw-loader!./examples/ExInline'
 
@@ -89,7 +99,9 @@
                 ExInline,
                 ExInlineCode,
                 ExEvents,
-                ExEventsCode
+                ExEventsCode,
+                ExMonth,
+                ExMonthCode
             }
         }
     }