+++ docs/pages/components/datepicker/examples/ExMonth.vue
@@ -1,0 +1,9 @@
+<template>
+    <b-field label="Select a date">
+        <b-datepicker
+            type="month"
+            placeholder="Click to select..."
+            icon="calendar-today">
+        </b-datepicker>
+    </b-field>
+</template>